moves getTPsInPattern for Element / Op to GenericSPARQLGraphPatternImpl1 / GenericSPARQLGraphPatternImpl2
